requirements.txt
Django==5.0.6
django-modeltranslation==0.19.5
djangorestframework==3.15.2
django-phonenumber-field==8.0.0
django-rosetta==0.10.0
django-jazzmin==3.0.0
python-dotenv==1.0.1
phonenumbers==8.13.40
drf-yasg==1.21.7
psycopg2-binary==2.9.9
setuptools==70.3.0
black==24.4.2
flake8==7.1.0
isort==5.13.2
pre-commit==3.7.1
